# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  handler.SubscriptionRequest:
    properties:
      end_date:
        example: "2025-10-01T00:00:00Z"
        type: string
      price:
        example: 499
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: "2023-10-01T00:00:00Z"
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  handler.SuccessResponse:
    properties:
      status:
        example: success
        type: string
    type: object
  handler.SummeryResponse:
    properties:
      total_price:
        example: 1497
        type: integer
    type: object
  model.Subscription:
    properties:
      end_date:
        example: "2024-10-01T00:00:00Z"
        type: string
      price:
        example: 299
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: "2023-10-01T00:00:00Z"
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
info:
  contact: {}
  description: REST API для управления онлайн-подписками и агрегации стоимости.
  title: SubService API
  version: "1.0"
paths:
  /subscriptions:
    delete:
      description: Удаляет запись о подписке по user_id и service_name
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        required: true
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success'
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: invalid user_id parameter / service_name is required
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: subscription not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      description: Возвращает одну подписку по user_id и service_name
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        required: true
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: subscription not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получить подписку
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создает запись о подписке пользователя
      parameters:
      - description: Данные подписки
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.SubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 'status: success'
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: invalid json / validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: subscription already exists
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Создать подписку
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновляет запись подписки (по user_id + service_name)
      parameters:
      - description: Данные подписки
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.SubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'status: success'
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: subscription not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/{userId}:
    get:
      description: Возвращает все подписки для пользователя
      parameters:
      - description: User ID (UUID)
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Subscription'
            type: array
        "400":
          description: invalid userId parameter
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Список подписок пользователя
      tags:
      - subscriptions
  /subscriptions/summary:
    get:
      description: Считает суммарную стоимость активных подписок по месяцам за период,
        с фильтрами
      parameters:
      - description: Начало периода (RFC3339)
        in: query
        name: from
        required: true
        type: string
      - description: Конец периода (RFC3339)
        in: query
        name: to
        required: true
        type: string
      - description: User ID (UUID)
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SummeryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Сумма подписок за период
      tags:
      - subscriptions
swagger: "2.0"